pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "studio-roundtime-monitor"
version = "1.0.0"
description = "Independent time monitoring module for SDP game systems"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Studio Development Team", email = "dev@studio.com"}
]
keywords = ["monitoring", "gaming", "time-analysis", "sdp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "asyncio-mqtt>=0.11.0",
    "sqlalchemy>=2.0",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "aiofiles>=23.0.0",
    "structlog>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0",
    "flake8>=6.0",
]
database = [
    "psycopg2-binary>=2.9.0",
    "pymysql>=1.0.0",
    "sqlite3",
]

[project.urls]
Homepage = "https://github.com/studio/studio-roundtime-monitor"
Documentation = "https://github.com/studio/studio-roundtime-monitor/docs"
Repository = "https://github.com/studio/studio-roundtime-monitor.git"
Issues = "https://github.com/studio/studio-roundtime-monitor/issues"

[project.scripts]
roundtime-monitor = "studio_roundtime_monitor.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["studio_roundtime_monitor*"]

[tool.setuptools.package-data]
studio_roundtime_monitor = ["config/*.yaml", "examples/*.py"]

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["studio_roundtime_monitor"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=studio_roundtime_monitor",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["studio_roundtime_monitor"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
